+++ test/test.sql
@@ -112,21 +112,21 @@ CREATE INDEX test_table_the_geom_webmercator_idx_p ON private_table USING gist (
 
 
 
-CREATE USER :PUBLICUSER WITH PASSWORD '';
-CREATE USER test_cartodb_user_1 WITH PASSWORD '';
+CREATE USER :PUBLICUSER WITH PASSWORD ':PUBLICPASS';
+CREATE USER :TESTUSER WITH PASSWORD ':TESTPASS';
 
-GRANT ALL ON TABLE untitle_table_4 TO test_cartodb_user_1;
+GRANT ALL ON TABLE untitle_table_4 TO :TESTUSER;
 GRANT SELECT ON TABLE untitle_table_4 TO :PUBLICUSER;
-GRANT ALL ON TABLE private_table TO test_cartodb_user_1;
-GRANT ALL ON SEQUENCE test_table_cartodb_id_seq_p TO test_cartodb_user_1;
+GRANT ALL ON TABLE private_table TO :TESTUSER;
+GRANT ALL ON SEQUENCE test_table_cartodb_id_seq_p TO :TESTUSER;
 
-GRANT ALL ON TABLE spatial_ref_sys TO test_cartodb_user_1, :PUBLICUSER;
+GRANT ALL ON TABLE spatial_ref_sys TO :TESTUSER, :PUBLICUSER;
 
 REVOKE ALL ON geometry_columns FROM public;
-GRANT ALL ON geometry_columns TO test_cartodb_user_1;
+GRANT ALL ON geometry_columns TO :TESTUSER;
 
 -- For https://github.com/CartoDB/CartoDB-SQL-API/issues/118
 DROP TABLE IF EXISTS cpg_test;
 CREATE TABLE cpg_test (a int);
-GRANT ALL ON TABLE cpg_test TO test_cartodb_user_1;
+GRANT ALL ON TABLE cpg_test TO :TESTUSER;
 GRANT SELECT ON TABLE cpg_test TO :PUBLICUSER;